backend/requirements.txt
flask
neo4j
python-dotenv
fernet
cryptography
